Refactor: Unified connector architecture
NEW STRUCTURE:
- lib/storage.js       → Storage class (was storage/index.js)
- lib/connector.js  → Base Connector class (was connectors/index.js)
- lib/remotes.js    → Git provider factory (was providers/index.js)
- lib/connectors/  → All external service connectors
  - github.js, gitlab.js, bitbucket.js, selfhosted.js, pinecone.js

All connectors now in one predictable place under lib/connectors/
Each connector in its own file - easy to discover/add new ones
