+++ lib/branch.js
@@ -15,7 +15,7 @@
  */
 
 // Try to use provider, fallback to CLI
-const { getProvider, detectProvider } = require('./providers');
+const { getProvider, detectProvider } = require('./remotes');
 
 let provider = null;
 let useProvider = false;
