+++ lib/connector.js
@@ -7,7 +7,7 @@
  *   const connector = getConnector('pinecone', { apiKey: 'xxx' });
  */
 
-const { VectorConnector, PineconeConnector } = require('./pinecone');
+const { VectorConnector, PineconeConnector } = require('./connectors/pinecone');
 
 function getConnector(type, config = {}) {
     switch (type) {
