+++ lib/connectors/pinecone.js
