+++ lib/remotes.js
@@ -1,0 +1,59 @@
+const { GitProvider } = require('./connectors/git');
+const { GitHubProvider } = require('./connectors/github');
+const { GitLabProvider } = require('./connectors/gitlab');
+const { BitbucketProvider } = require('./connectors/bitbucket');
+const { SelfHostedProvider } = require('./connectors/selfhosted');
+
+function detectProvider() {
+    try {
+        const { execSync } = require('child_process');
+        const remoteUrl = execSync('git remote get-url origin', { encoding: 'utf8' }).trim();
+
+        if (remoteUrl.includes('github.com') || remoteUrl.includes('github')) {
+            return 'github';
+        } else if (remoteUrl.includes('gitlab.com') || remoteUrl.includes('gitlab')) {
+            return 'gitlab';
+        } else if (remoteUrl.includes('bitbucket.org') || remoteUrl.includes('bitbucket')) {
+            return 'bitbucket';
+        } else {
+            return 'selfhosted';
+        }
+    } catch (e) {
+        return 'selfhosted';
+    }
+}
+
+function getProvider(type = null, config = {}) {
+    const providerType = type || detectProvider();
+    switch (providerType) {
+        case 'github':
+            return new GitHubProvider(config);
+        case 'gitlab':
+            return new GitLabProvider(config);
+        case 'bitbucket':
+            return new BitbucketProvider(config);
+        case 'selfhosted':
+        default:
+            return new SelfHostedProvider(config);
+    }
+}
+
+function getAllProviders(config = {}) {
+    return {
+        github: new GitHubProvider(config),
+        gitlab: new GitLabProvider(config),
+        bitbucket: new BitbucketProvider(config),
+        selfhosted: new SelfHostedProvider(config)
+    };
+}
+
+module.exports = {
+    getProvider,
+    detectProvider,
+    getAllProviders,
+    GitProvider,
+    GitHubProvider,
+    GitLabProvider,
+    BitbucketProvider,
+    SelfHostedProvider
+};
