+++ lib/storage.js
