+++ lib/sync.js
@@ -158,7 +158,7 @@ function getAllCircuits() {
  * @returns {GitProvider[]} Array of configured providers
  */
 function getConfiguredProviders() {
-    const { getAllProviders } = require('./providers');
+    const { getAllProviders } = require('./remotes');
     const providers = getAllProviders();
     
     const configured = [];
@@ -341,7 +341,7 @@ function isRAID() {
  * @returns {object} Rebase result
  */
 async function rebase(providerName) {
-    const { getAllProviders } = require('./providers');
+    const { getAllProviders } = require('./remotes');
     const providers = getAllProviders();
     const provider = providers[providerName];
